+++ iam/iam-for-deploy-bot/main.tf
@@ -5,6 +5,10 @@ resource "aws_iam_user" "user" {
   }
 }
 
+output "user_name" {
+  value = aws_iam_user.user.name
+}
+
 resource "aws_iam_user_group_membership" "memberships" {
   user   = "deploy-bot"
   groups = ["deploy-bot-deploy-access"]
@@ -127,6 +131,7 @@ resource "aws_iam_policy" "deploy-bot-extra-access" {
           "ecs:ListServices",
           "ecs:ListTasks",
           "ecs:DescribeTaskDefinition",
+          "ecs:TagResource",
           "elasticloadbalancing:Describe*",
           "iam:GetRole",
           "logs:*",