+++ rds/variables.tf
@@ -26,8 +26,9 @@ variable "instance_class" {
 }
 
 variable "engine_version" {
-  default = "16"
-  type    = string
+  default  = "16"
+  type     = string
+  nullable = false
 }
 
 variable "allocated_storage" {